+++ src/request.ts
@@ -34,7 +34,10 @@ async function addRequest(
     let batchQuery = batchQueries.get(endpoint)
     if (!batchQuery) {
         // init new batch
-        let timerHandle = window.setTimeout(() => batchRequest(endpoint, batchQueries.get(endpoint)), batchTimeout)
+        let timerHandle = window.setTimeout(() => {
+            batchRequest(endpoint, batchQueries.get(endpoint));
+            batchQueries.delete(endpoint);
+        }, batchTimeout)
         batchQuery = {
             timerHandle,
             queries: [],
@@ -47,7 +50,9 @@ async function addRequest(
     batchQuery.onload.push(onload);
     batchQuery.onerror.push(onerror);
     // send or save
+    console.debug(`Batch query size: ${batchQuery.queries.length}`);
     if (batchQuery.queries.length >= maxBatchSize) {
+        console.debug(`Batch query full: Run it.`)
         window.clearTimeout(batchQuery.timerHandle);
         batchQueries.delete(endpoint);
         return batchRequest(endpoint, batchQuery);